Fix a silent int16 overflow in the Python plane stage
The workload sweep's equivalence gate caught this, which is exactly what the
gate is for. python/grasp_core/plane.py accumulated each block's inlier count
with dtype=np.int16 under a comment asserting that a block cannot overflow it.
A block holds ransac_block_points rows, so anything above 32767 wraps. When the
sweep's block calibration reached 131072 points the counts went negative,
RANSAC selected a different candidate from C++, and the two implementations
disagreed by 0.139 in the plane normal and by 0.0395 rad in the joint solution,
with graspable flipping on two frames.

assets/pipeline_config.json claimed of that key that any value gives
bit-identical output. That claim was false and is now both corrected and
tested: python/tests/test_block_size_invariance.py runs the pipeline at 512,
4096, 32767, 32768 and 131072 points per block and asserts the plane, the
graspable flag and the joint solution are bit-identical, not merely close.
Verified the test catches the original bug by restoring int16, which fails it.

int32 rather than int64 keeps the narrow-accumulator speedup the comment was
after while removing the overflow.

+++ python/grasp_core/plane.py
@@ -53,7 +53,7 @@ class PlaneRemover:
         self._hit_flat = np.empty(block_points * k, dtype=np.bool_)
         self._plane_flat = np.empty(4 * k, dtype=np.float64)
         self._counts = np.empty(k, dtype=np.int64)
-        self._partial = np.empty(k, dtype=np.int16)
+        self._partial = np.empty(k, dtype=np.int32)
         self._normals = np.empty((k, 3), dtype=np.float64)
         self._offsets = np.empty(k, dtype=np.float64)
         self._valid = np.empty(k, dtype=np.bool_)
@@ -137,10 +137,13 @@ class PlaneRemover:
             np.matmul(homogeneous[start:stop], model, out=chunk)
             np.abs(chunk, out=chunk)
             np.less(chunk, threshold, out=inside)
-            # A bool view summed as int16 is three times cheaper than the
-            # int64 reduction numpy picks by default, and a block cannot
-            # overflow it.
-            np.add.reduce(inside.view(np.uint8), axis=0, dtype=np.int16,
+            # A bool view summed in a narrow integer type is markedly cheaper
+            # than the int64 reduction numpy picks by default. int32, not
+            # int16: the sum reaches the block's row count, so int16 wraps
+            # silently for any ransac_block_points above 32767 and RANSAC then
+            # picks a different candidate. That is not hypothetical, it was
+            # caught by the equivalence gate at a 131072-point block.
+            np.add.reduce(inside.view(np.uint8), axis=0, dtype=np.int32,
                           out=partial)
             np.add(counts, partial, out=counts)
